Move contents of SiPixelGenErrorDBObjectESProducer header to source file
